CI: add initial support of Deno build
Motivation #2566
