+++ .babelrc-deno.json
@@ -1,0 +1,13 @@
+{
+  "plugins": [
+    "@babel/plugin-transform-flow-strip-types",
+    ["./resources/add-extension-to-import-paths", { "extension": "mjs" }],
+    "./resources/inline-invariant"
+  ],
+  "overrides": [
+    {
+      "include": "src/error/GraphQLError.js",
+      "plugins": [["@babel/plugin-transform-classes", { "loose": false }]]
+    }
+  ]
+}
