+++ .github/workflows/ci.yml
@@ -49,8 +49,11 @@ jobs:
       - name: Spellcheck
         run: npm run check:spelling
 
-      - name: Build package
-        run: npm run build
+      - name: Build NPM package
+        run: npm run build:npm
+
+      - name: Build Deno package
+        run: npm run build:deno
 
   fuzz:
     name: Run fuzzing tests
@@ -196,10 +199,46 @@ jobs:
       - name: Install Dependencies
         run: npm ci
 
-      - name: Build package
-        run: npm run build
+      - name: Build NPM package
+        run: npm run build:npm
 
       - name: Deploy to `npm` branch
-        run: npm run gitpublish
+        run: npm run gitpublish:npm
+        env:
+          GH_TOKEN: ${{ secrets.GH_NPM_BRANCH_PUBLISH_TOKEN }}
+
+  deploy-to-deno-branch:
+    name: Deploy to `deno` branch
+    runs-on: ubuntu-latest
+    if: |
+      github.event_name == 'push' &&
+      github.repository == 'graphql/graphql-js' &&
+      github.ref == 'refs/heads/master'
+    needs: [test, fuzz, lint]
+    steps:
+      - name: Checkout repo
+        uses: actions/checkout@v2
+
+      - name: Setup Node.js
+        uses: actions/setup-node@v1
+        with:
+          node-version: ${{ env.NODE_VERSION_USED_FOR_DEVELOPMENT }}
+
+      - name: Cache Node.js modules
+        uses: actions/cache@v2
+        with:
+          path: ~/.npm
+          key: ${{ runner.OS }}-node-${{ hashFiles('**/package-lock.json') }}
+          restore-keys: |
+            ${{ runner.OS }}-node-
+
+      - name: Install Dependencies
+        run: npm ci
+
+      - name: Build Deno package
+        run: npm run build:deno
+
+      - name: Deploy to `npm` branch
+        run: npm run gitpublish:deno
         env:
           GH_TOKEN: ${{ secrets.GH_NPM_BRANCH_PUBLISH_TOKEN }}
