+++ resources/gitpublish.sh
@@ -1,43 +1,47 @@
-#!/bin/sh
+#!/bin/bash
+
+#  This script maintains a git branch which mirrors master but in a form that
+#  what will eventually be deployed to npm, allowing npm dependencies to use:
+#
+#      "graphql": "git://github.com/graphql/graphql-js.git#npm"
+#
+#  Additionaly it use use to push Deno build to `deno` branch.
+
+BRANCH=$1
+DIST_DIR=$2
 
 # Exit immediately if any subcommand terminated
-trap "exit 1" ERR
+set -e
 
-# This script maintains a git branch which mirrors master but in a form that
-# what will eventually be deployed to npm, allowing npm dependencies to use:
-#
-#     "graphql": "git://github.com/graphql/graphql-js.git#npm"
-#
-
-if [ ! -d "./dist" ]; then
- echo 'Directory `dist` does not  exist, please run `npm run build`!'
+if [ ! -d $DIST_DIR ]; then
+ echo "Directory `${DIST_DIR}` does not  exist!"
  exit 1;
 fi;
 
-# Create empty npm directory
-rm -rf npm
-git clone -b npm "https://${GH_TOKEN}@github.com/graphql/graphql-js.git" npm
+# Create empty directory
+rm -rf $BRANCH
+git clone -b $BRANCH "https://${GH_TOKEN}@github.com/graphql/graphql-js.git" $BRANCH
 
 # Remove existing files first
-rm -rf npm/**/*
-rm -rf npm/*
+rm -rf $BRANCH/**/*
+rm -rf $BRANCH/*
 
 # Copy over necessary files
-cp -r dist/* npm/
+cp -r $DIST_DIR/* $BRANCH/
 
 # Reference current commit
 HEADREV=`git rev-parse HEAD`
 echo $HEADREV
 
 # Deploy
-cd npm
+cd $BRANCH
 git config user.name "GitHub Action Script"
 git config user.email "please@open.issue"
 git add -A .
 if git diff --staged --quiet; then
   echo "Nothing to publish"
 else
-  git commit -a -m "Deploy $HEADREV to NPM branch"
+  git commit -a -m "Deploy $HEADREV to `$BRANCH` branch"
   git push > /dev/null 2>&1
   echo "Pushed"
 fi
